Reorganize i18n language list and utils
