+++ site/gatsby-site/config.js
@@ -143,12 +143,12 @@ const config = {
     translationsConnectionString: process.env.MONGODB_TRANSLATIONS_CONNECTION_STRING,
   },
   google: {
-    translateApikey: process.env.GOOGLE_TRANSLATE_API_KEY,
-    availableLanguages: process.env.GATSBY_AVAILABLE_LANGUAGES || 'en',
     mapsApiKey: process.env.GOOGLE_MAPS_API_KEY,
   },
   i18n: {
-    locales: process.env.GATSBY_AVAILABLE_LANGUAGES,
+    availableLanguages: String(process.env.GATSBY_AVAILABLE_LANGUAGES).split(',') || ['en'],
+    translateApikey: process.env.GOOGLE_TRANSLATE_API_KEY,
+    defaultLocale: 'en',
   },
 };
